requirements.txt
streamlit
gspread>=5.7.2,<5.8.0
google-auth>=2.22.0,<2.23.0
boto3==1.26.165
snowflake-snowpark-python>=1.9.0